Restructure table-driven tests to the proposed way; it's much cleaner this way
